+++ src/core/io/lottie/tgs_format.cpp
@@ -9,6 +9,7 @@
 #include "model/shapes/stroke.hpp"
 #include "model/shapes/repeater.hpp"
 #include "model/shapes/inflate_deflate.hpp"
+#include "model/shapes/offset_path.hpp"
 #include "validation.hpp"
 
 
@@ -60,6 +61,10 @@ private:
         else if ( qobject_cast<model::InflateDeflate*>(node) )
         {
             show_error(node, TgsFormat::tr("Inflate/Deflate is not supported"), app::log::Warning);
+        }
+        else if ( qobject_cast<model::OffsetPath*>(node) )
+        {
+            show_error(node, TgsFormat::tr("Offset Path is not supported"), app::log::Warning);
         }
     }
 };