+++ src/core/model/shapes/stroke.hpp
@@ -33,7 +33,6 @@ private:
     Q_ENUM(Cap);
     Q_ENUM(Join);
 
-
     GLAXNIMATE_ANIMATABLE(float, width, 1, {}, 0)
     GLAXNIMATE_PROPERTY(Cap, cap, RoundCap, nullptr, nullptr, PropertyTraits::Visual)
     GLAXNIMATE_PROPERTY(Join, join, RoundJoin, nullptr, nullptr, PropertyTraits::Visual)