+++ src/gui/widgets/dialogs/gw_impl_ui.cpp
@@ -16,6 +16,7 @@
 #include "model/shapes/trim.hpp"
 #include "model/shapes/inflate_deflate.hpp"
 #include "model/shapes/round_corners.hpp"
+#include "model/shapes/offset_path.hpp"
 
 #include "widgets/dialogs/io_status_dialog.hpp"
 #include "widgets/dialogs/about_dialog.hpp"
@@ -199,6 +200,7 @@ void GlaxnimateWindow::Private::init_actions()
     add_modifier_menu_action<model::Trim>(ui.menu_new_layer);
     add_modifier_menu_action<model::InflateDeflate>(ui.menu_new_layer);
     add_modifier_menu_action<model::RoundCorners>(ui.menu_new_layer);
+    add_modifier_menu_action<model::OffsetPath>(ui.menu_new_layer);
     connect(ui.action_import_image, &QAction::triggered, parent, [this]{import_image();});
     connect(ui.action_delete, &QAction::triggered, parent, &GlaxnimateWindow::delete_selected);
     connect(ui.action_export, &QAction::triggered, parent, &GlaxnimateWindow::document_export);