+++ src/gui/widgets/node_menu.cpp
@@ -13,6 +13,7 @@
 #include "model/shapes/text.hpp"
 #include "model/shapes/inflate_deflate.hpp"
 #include "model/shapes/round_corners.hpp"
+#include "model/shapes/offset_path.hpp"
 
 #include "model/assets/assets.hpp"
 
@@ -221,6 +222,7 @@ void actions_group(QMenu* menu, GlaxnimateWindow* window, model::Group* group)
     add_child_action<model::Repeater>(menu_add, group);
     add_child_action<model::InflateDeflate>(menu_add, group);
     add_child_action<model::RoundCorners>(menu_add, group);
+    add_child_action<model::OffsetPath>(menu_add, group);
 
     menu->addSeparator();